model cleanup

+++ lib/ain-ocean/src/model/oracle_price_active.rs
@@ -21,14 +21,16 @@ pub struct OraclePriceActive {
 #[serde(rename_all = "camelCase")]
 pub struct OraclePriceActiveNext {
     #[serde(with = "rust_decimal::serde::str")]
-    pub amount: Decimal,
-    pub weightage: u32,
+    pub amount: Decimal, // aggregated_amount
+    #[serde(with = "rust_decimal::serde::str")]
+    pub weightage: Decimal, // aggregated_weightage
     pub oracles: OraclePriceActiveNextOracles,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone, Default)]
 #[serde(rename_all = "camelCase")]
 pub struct OraclePriceActiveNextOracles {
-    pub active: i32,
+    #[serde(with = "rust_decimal::serde::str")]
+    pub active: Decimal, // aggregated_count
     pub total: i32,
 }
